+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## Change
 * predict with type "ratingMatrix" now returns predictions for the known ratings instead of replacing them with the known values.
-* Recommender methods Popular and AR now also return ratings for binary data (and thus can be used for HybridRecommender).
+* Recommender methods Popular, AR and RERECOMMENDER now also return ratings for binary data (and thus can be used for HybridRecommender).
 
 ## Bugfixes
 * evaluationScheme with negative numbers for given (all-but-x scheme) now works even if there are not given items left (reported by philippschmalen).